Just testing the app with a function in MoviesScreen

+++ lib/movies/data/models/movie_model.dart
@@ -18,7 +18,7 @@ class MovieModel extends MovieEntity {
       backdropPath: jsonData['poster_path'],
       genreIds: List<int>.from(jsonData['genre_ids'].map((e) => e)),
       overview: jsonData['overview'],
-      voteAverage: jsonData['vote_average'],
+      voteAverage: jsonData['vote_average'].toDouble(),
       releaseDate: jsonData['release_date'],
     );
   }
